requirements.txt
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
black==25.12.0
bleach==6.3.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
fonttools==4.61.1
idna==3.11
imbalanced-learn==0.14.1
imblearn==0.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kaggle==1.8.3
kagglehub==0.3.13
kagglesdk==0.1.14
kiwisolver==1.4.9
librt==0.7.5
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.14.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
patsy==1.0.2
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
reportlab==4.4.7
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.5
stack-data==0.6.3
statsmodels==0.14.6
text-unidecode==1.3
threadpoolctl==3.6.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.4
